% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshot}
\alias{snapshot}
\title{Take a snapshot of some variable's values or compare the values to a previously taken snapshot.}
\usage{
snapshot(...)
}
\description{
This function has two modes, (1) record, and (2) check.
Depending on the result of function 'refcmpEnv$is_reference_run', the mode is (1) if TRUE is returned or (2) otherwise.
}
\details{
When in the first mode 'record', function 'snapshot' serializes the values of the specified variables into a file together with the names of the variables. It also increases the ID of the snapshot automatically.

When in the second mode 'check', function 'snapshot' deserializes values of a previously taken snapshot and compares the values of the provided variables to the deserialized values by matching names.
}
\examples{
x <- 10
y <- 20
z <- function() print("hello")

snapshot(x, y, z)
}
