% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snapshot.show}
\alias{snapshot.show}
\title{Shows the contents of a snapshot.}
\usage{
snapshot.show(id = refcmpEnv$snapshot_id)
}
\arguments{
\item{id}{The ID of the snapshot to show (default: latest)}
}
\description{
Shows the contents of a snapshot.
}
\examples{
a <- 1
snapshot(a)
snapshot.show()

a <- 2
snapshot(a)

a <- 3
snapshot(a)

a <- 4
snapshot(a)
snapshot.show(0)
snapshot.show(1)
snapshot.show(2)
snapshot.show(3)
}
